uc-crux-llvm: Hide PartType invariants behind module boundary
Refactoring only, no functional change. Enforces API invariant with careful
export list.
