+++ uc-crux-llvm/src/UCCrux/LLVM/Classify.hs
@@ -69,9 +69,8 @@ import           UCCrux.LLVM.Context.Module (ModuleContext, dataLayout)
 import           UCCrux.LLVM.Context.Function (FunctionContext, argumentNames, moduleTypes)
 import           UCCrux.LLVM.Constraints
 import           UCCrux.LLVM.Cursor (ppCursor, Selector(..), SomeInSelector(SomeInSelector))
-import           UCCrux.LLVM.FullType (MapToCrucibleType, IsPtrRepr(..), isPtrRepr, FullTypeRepr(..), PartTypeRepr)
-import           UCCrux.LLVM.FullType.MemType (toMemType, asFullType)
-import           UCCrux.LLVM.FullType.ModuleTypes (ModuleTypes)
+import           UCCrux.LLVM.FullType (MapToCrucibleType, IsPtrRepr(..), isPtrRepr, FullTypeRepr(..), PartTypeRepr, ModuleTypes, asFullType)
+import           UCCrux.LLVM.FullType.MemType (toMemType)
 import           UCCrux.LLVM.Logging (Verbosity(Hi))
 import           UCCrux.LLVM.Setup (SymValue)
 import           UCCrux.LLVM.Setup.Monad (TypedSelector(..))
